requirements-plant-model.txt
torch
torchvision
transformers
pillow
pyserial
requests
opencv-python
numpy
